requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
lxml==5.3.0
multidict==6.1.0
openai==1.55.0
packaging==24.2
pip-review==1.3.0
propcache==0.2.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.8.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
